feat(drivers): add memory driver
Added memfs which provides a module that has an API identical to node's `fs` but operates on an in
memory filesystem.
