+++ package.json
@@ -24,7 +24,11 @@
     "dependencies": {
         "aws-sdk": "^2.431.0",
         "ramda": "0.25.0",
-        "stream-to-array": "^2.3.0"
+        "stream-to-array": "^2.3.0",
+        "verror": "^1.10.0"
+    },
+    "peerDependencies": {
+        "memfs": "^2.15.2"
     },
     "devDependencies": {
         "@commitlint/cli": "^7.5.2",
@@ -32,6 +36,7 @@
         "@types/jest": "^24.0.11",
         "@types/ramda": "types/npm-ramda#dist",
         "@types/stream-to-array": "^2.3.0",
+        "@types/verror": "^1.10.3",
         "@typescript-eslint/eslint-plugin": "^1.5.0",
         "@typescript-eslint/parser": "^1.5.0",
         "commitizen": "^3.0.7",
