+++ tsconfig.json
@@ -5,6 +5,6 @@
         "declaration": true,
         "outDir": "./dist",
         "strict": true,
-        "downlevelIteration": true,
+        "downlevelIteration": true
     }
 }
